refactor: RK into separate module.  More broadcasting compatibility
